fix(presentation): hoist the entity renderer instead of copying it
Review fixes for #2155: the 180-line copy of compact_entity_json in
impl_vcs is replaced by a neutral crate-root entity_render module that
both the engine and presentation import (presentation re-exports it, so
its surface is unchanged). document_body and document_migration move
onto the keystone conversion, closing the transitive leak the fence
could not see, and the fence now also scans the crate-root engine
modules (transport modules stay out of scope by design).

Refs #2155
