+++ crates/reddb-server/src/runtime/impl_vcs.rs
@@ -1700,186 +1700,6 @@ impl RedDBRuntime {
 /// Phase 6 will stage the merged body (base/ours/theirs payloads +
 /// JSON 3-way merge result) so `vcs_conflict_resolve` can apply the
 /// resolved value back to the user collection.
-fn compact_vcs_entity_json(entity: &UnifiedEntity) -> JsonValue {
-    let mut object = Map::new();
-    object.insert("id".to_string(), JsonValue::Number(entity.id.raw() as f64));
-    object.insert(
-        "kind".to_string(),
-        JsonValue::String(entity.kind.storage_type().to_string()),
-    );
-    object.insert(
-        "collection".to_string(),
-        JsonValue::String(entity.kind.collection().to_string()),
-    );
-    object.insert(
-        "red_entity_type".to_string(),
-        JsonValue::String(vcs_entity_type(entity).to_string()),
-    );
-    object.insert(
-        "red_capabilities".to_string(),
-        JsonValue::Array(
-            vcs_entity_capabilities(entity)
-                .into_iter()
-                .map(JsonValue::String)
-                .collect(),
-        ),
-    );
-
-    match &entity.data {
-        EntityData::Row(row) => {
-            if let Some(named) = &row.named {
-                object.insert(
-                    "row".to_string(),
-                    JsonValue::Object(vcs_named_fields_json(named)),
-                );
-            }
-        }
-        EntityData::Node(node) => {
-            object.insert(
-                "properties".to_string(),
-                JsonValue::Object(
-                    node.properties
-                        .iter()
-                        .map(|(key, value)| (key.clone(), value.to_json()))
-                        .collect(),
-                ),
-            );
-        }
-        EntityData::Edge(edge) => {
-            object.insert(
-                "weight".to_string(),
-                JsonValue::Number(f64::from(edge.weight)),
-            );
-            object.insert(
-                "properties".to_string(),
-                JsonValue::Object(
-                    edge.properties
-                        .iter()
-                        .map(|(key, value)| (key.clone(), value.to_json()))
-                        .collect(),
-                ),
-            );
-        }
-        EntityData::Vector(vector) => {
-            object.insert(
-                "dense".to_string(),
-                JsonValue::Array(
-                    vector
-                        .dense
-                        .iter()
-                        .map(|value| JsonValue::Number(f64::from(*value)))
-                        .collect(),
-                ),
-            );
-            if let Some(content) = &vector.content {
-                object.insert("content".to_string(), JsonValue::String(content.clone()));
-            }
-        }
-        EntityData::TimeSeries(timeseries) => {
-            object.insert(
-                "metric".to_string(),
-                JsonValue::String(timeseries.metric.clone()),
-            );
-            object.insert(
-                "timestamp_ns".to_string(),
-                JsonValue::Number(timeseries.timestamp_ns as f64),
-            );
-            object.insert("value".to_string(), JsonValue::Number(timeseries.value));
-        }
-        EntityData::QueueMessage(message) => {
-            object.insert("payload".to_string(), message.payload.to_json());
-            object.insert(
-                "attempts".to_string(),
-                JsonValue::Number(f64::from(message.attempts)),
-            );
-            object.insert("acked".to_string(), JsonValue::Bool(message.acked));
-        }
-    }
-
-    JsonValue::Object(object)
-}
-
-fn vcs_entity_type(entity: &UnifiedEntity) -> &'static str {
-    match (&entity.kind, &entity.data) {
-        (EntityKind::TableRow { .. }, EntityData::Row(row)) if vcs_row_is_kv(row) => "kv",
-        (EntityKind::TableRow { .. }, EntityData::Row(_)) => "table",
-        (EntityKind::GraphNode(_), EntityData::Node(_)) => "graph_node",
-        (EntityKind::GraphEdge(_), EntityData::Edge(_)) => "graph_edge",
-        (EntityKind::Vector { .. }, EntityData::Vector(_)) => "vector",
-        (EntityKind::TimeSeriesPoint(_), EntityData::TimeSeries(_)) => "timeseries",
-        _ => "unknown",
-    }
-}
-
-fn vcs_entity_capabilities(entity: &UnifiedEntity) -> Vec<String> {
-    let capabilities = match (&entity.kind, &entity.data) {
-        (EntityKind::TableRow { .. }, EntityData::Row(row)) => {
-            let mut values = BTreeSet::from(["table".to_string(), "structured".to_string()]);
-            if vcs_row_is_kv(row) {
-                values.insert("kv".to_string());
-            }
-            if row
-                .named
-                .as_ref()
-                .is_some_and(|named| named.values().any(vcs_documentish_value))
-                || row.columns.iter().any(vcs_documentish_value)
-            {
-                values.insert("document".to_string());
-            }
-            values
-        }
-        (EntityKind::GraphNode(_), EntityData::Node(_)) => {
-            BTreeSet::from(["graph".to_string(), "graph_node".to_string()])
-        }
-        (EntityKind::GraphEdge(_), EntityData::Edge(_)) => {
-            BTreeSet::from(["graph".to_string(), "graph_edge".to_string()])
-        }
-        (EntityKind::Vector { .. }, EntityData::Vector(_)) => BTreeSet::from([
-            "vector".to_string(),
-            "similarity".to_string(),
-            "embedding".to_string(),
-        ]),
-        (EntityKind::TimeSeriesPoint(_), EntityData::TimeSeries(_)) => BTreeSet::from([
-            "document".to_string(),
-            "timeseries".to_string(),
-            "metric".to_string(),
-            "temporal".to_string(),
-        ]),
-        _ => BTreeSet::new(),
-    };
-    capabilities.into_iter().collect()
-}
-
-fn vcs_documentish_value(value: &Value) -> bool {
-    matches!(value, Value::Json(_) | Value::Blob(_))
-}
-
-fn vcs_row_is_kv(row: &RowData) -> bool {
-    let Some(named) = row.named.as_ref() else {
-        return false;
-    };
-    match named.len() {
-        1 => named.contains_key("key") || named.contains_key("value"),
-        2 => named.contains_key("key") && named.contains_key("value"),
-        _ => false,
-    }
-}
-
-fn vcs_named_fields_json(named: &HashMap<String, Value>) -> Map<String, JsonValue> {
-    let mut object = named
-        .iter()
-        .map(|(key, value)| (key.clone(), value.to_json()))
-        .collect::<Map<_, _>>();
-    if let Some(Value::Json(bytes)) = named.get("body") {
-        if let Some(fields) = crate::document_body::body_fields(bytes) {
-            for (key, value) in fields {
-                object.entry(key).or_insert_with(|| value.to_json());
-            }
-        }
-    }
-    object
-}
-
 fn materialize_merge_conflicts(
     rt: &RedDBRuntime,
     store: &UnifiedStore,
@@ -1926,7 +1746,7 @@ fn materialize_merge_conflicts(
                 continue;
             }
             let eid = entity.id.raw();
-            let body = compact_vcs_entity_json(&entity);
+            let body = crate::entity_render::compact_entity_json(&entity);
             if base_snap.sees(xmin, xmax) {
                 at_base.insert(eid, body.clone());
             }
