+++ crates/reddb-server/tests/engine_presentation_boundary.rs
@@ -3,7 +3,9 @@ use std::path::{Path, PathBuf};
 
 fn rust_files(root: &Path, files: &mut Vec<PathBuf>) {
     for entry in fs::read_dir(root).expect("engine source directory should be readable") {
-        let path = entry.expect("engine source entry should be readable").path();
+        let path = entry
+            .expect("engine source entry should be readable")
+            .path();
         if path.is_dir() {
             rust_files(&path, files);
         } else if path.extension().is_some_and(|extension| extension == "rs") {
@@ -17,9 +19,28 @@ fn engine_layers_do_not_import_presentation() {
     let source = Path::new(env!("CARGO_MANIFEST_DIR")).join("src");
     let mut violations = Vec::new();
 
-    for layer in ["runtime", "storage", "application"] {
+    // Crate-root modules that engine layers consume count too: a
+    // presentation import in `document_body.rs` (used by 16 runtime/storage/
+    // application files) is the same leak wearing a different path (issue
+    // #2155 review). Transport modules at the root (server.rs, rpc_stdio.rs)
+    // are presentation consumers by design and stay out of scope.
+    const ROOT_ENGINE_MODULES: &[&str] = &[
+        "document_body.rs",
+        "document_migration.rs",
+        "entity_render.rs",
+    ];
+
+    for (layer, root_modules) in [
+        ("runtime", None),
+        ("storage", None),
+        ("application", None),
+        ("<crate root>", Some(ROOT_ENGINE_MODULES)),
+    ] {
         let mut files = Vec::new();
-        rust_files(&source.join(layer), &mut files);
+        match root_modules {
+            Some(modules) => files.extend(modules.iter().map(|module| source.join(module))),
+            None => rust_files(&source.join(layer), &mut files),
+        }
 
         for file in files {
             let contents = fs::read_to_string(&file).expect("engine source should be readable");
